Create README.md
📝 Description:

This is a personal portfolio website designed and developed to showcase my skills, projects, and contact information. The main goal of this website is to introduce myself as a web developer and provide visitors (like potential clients or employers) with a clear and professional view of my work and abilities.